Remove duplicate file with wrong spelling
